backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==1.10.15
python-multipart==0.0.9
ultralytics==8.1.41
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
websockets==12.0
